[storeFileUploadStatusChangesInNewCollection] Rename repo
