+++ app/FrontendModule.scala
@@ -20,8 +20,6 @@ import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.http._
 import uk.gov.hmrc.play.http.ws.UploadHttpClient
 import uk.gov.hmrc.uploaddocuments.connectors.FrontendAuthConnector
-import uk.gov.hmrc.uploaddocuments.repository.{CacheRepository, JourneyCacheRepository}
-import uk.gov.hmrc.uploaddocuments.services._
 
 class FrontendModule(val environment: Environment, val configuration: Configuration) extends AbstractModule {
   override def configure(): Unit = {
