+++ app/uk/gov/hmrc/uploaddocuments/controllers/internal/CallbackFromUpscanController.scala
@@ -22,7 +22,7 @@ import uk.gov.hmrc.uploaddocuments.connectors.FileUploadResultPushConnector
 import uk.gov.hmrc.uploaddocuments.controllers.{BaseController, BaseControllerComponents}
 import uk.gov.hmrc.uploaddocuments.journeys.JourneyModel.canOverwriteFileUploadStatus
 import uk.gov.hmrc.uploaddocuments.models._
-import uk.gov.hmrc.uploaddocuments.repository.NewJourneyCacheRepository.DataKeys
+import uk.gov.hmrc.uploaddocuments.repository.JourneyCacheRepository.DataKeys
 import uk.gov.hmrc.uploaddocuments.support.UploadLog
 import uk.gov.hmrc.uploaddocuments.utils.LoggerUtil
